- warnings: Fix some warnings reported by clang (mostly unused member variables)
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@10636 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/infowindow.cpp
@@ -129,7 +129,7 @@ class Stacker
 static Stacker stacker;
 
 static std::list<wxString> active_messages;
-static unsigned lastDisplay = std::numeric_limits<long>::max();
+static unsigned lastDisplay = std::numeric_limits<unsigned>::max();
 static wxRect displayGeometry;
 }
